EShop-DDD/Test/Domain.Test/Aggregates/Users/ValueObjects/RoleUnitTest.cs
using Domain.Aggregates.Users.ValueObjects;
using Domain.SharedKernel;

namespace Domain.Test.Aggregates.Users.ValueObjects
{
	public class RoleUnitTest : object
	{
		public RoleUnitTest() : base()
		{
		}

		[Fact]
		public void Add_Null_value_In_Role()
		{
			var result =
				Role.GetByValue(value: null);

			// **************************************************
			Xunit.Assert.True(condition: result.IsFailed);
			Xunit.Assert.False(condition: result.IsSuccess);
			// **************************************************

			// **************************************************
			string errorMessage = string.Format
				(Resources.Messages.Validations.Required,
				Resources.DataDictionary.UserRole);

			Xunit.Assert.Equal
				(expected: errorMessage, actual: result.Errors[0].Message);
			// **************************************************

			// **************************************************
			Xunit.Assert.Single(result.Errors);
			Xunit.Assert.Empty(result.Successes);
			// **************************************************
		}

		[Xunit.Fact]
		public void Add_Out_Of_Range_Value_In_Role()
		{
			var result =
				Role.GetByValue(value: 10);

			// **************************************************
			Xunit.Assert.True(condition: result.IsFailed);
			Xunit.Assert.False(condition: result.IsSuccess);
			// **************************************************

			// **************************************************
			string errorMessage = string.Format
				(Resources.Messages.Validations.InvalidCode,
				Resources.DataDictionary.UserRole);

			Xunit.Assert.Equal
				(expected: errorMessage, actual: result.Errors[0].Message);
			// **************************************************

			// **************************************************
			Xunit.Assert.Single(result.Errors);
			Xunit.Assert.Empty(result.Successes);
			// **************************************************
		}

		[Xunit.Fact]
		public void Add_Customer_Role()
		{
			var result =
				Role.GetByValue(value: 0);

			// **************************************************
			Xunit.Assert.False(condition: result.IsFailed);
			Xunit.Assert.True(condition: result.IsSuccess);
			// **************************************************

			// **************************************************
			Xunit.Assert.Equal
				(expected: Role.Customer,
				actual: result.Value);
			// **************************************************

			// **************************************************
			Xunit.Assert.Empty(result.Errors);
			Xunit.Assert.Empty(result.Successes);
			// **************************************************
		}

		[Xunit.Fact]
		public void Add_Agent_Role()
		{
			var result =
				Role.GetByValue(value: 1);

			// **************************************************
			Xunit.Assert.False(condition: result.IsFailed);
			Xunit.Assert.True(condition: result.IsSuccess);
			// **************************************************

			// **************************************************
			Xunit.Assert.Equal
				(expected: Role.Agent,
				actual: result.Value);
			// **************************************************

			// **************************************************
			Xunit.Assert.Empty(result.Errors);
			Xunit.Assert.Empty(result.Successes);
			// **************************************************
		}

		[Xunit.Fact]
		public void Add_Supervisor_Role()
		{
			var result =
				Role.GetByValue(value: 2);

			// **************************************************
			Xunit.Assert.False(condition: result.IsFailed);
			Xunit.Assert.True(condition: result.IsSuccess);
			// **************************************************

			// **************************************************
			Xunit.Assert.Equal
				(expected: Role.Supervisor,
				actual: result.Value);
			// **************************************************

			// **************************************************
			Xunit.Assert.Empty(result.Errors);
			Xunit.Assert.Empty(result.Successes);
			// **************************************************
		}

		[Xunit.Fact]
		public void Add_Administrator_Role()
		{
			var result =
				Role.GetByValue(value: 3);

			// **************************************************
			Xunit.Assert.False(condition: result.IsFailed);
			Xunit.Assert.True(condition: result.IsSuccess);
			// **************************************************

			// **************************************************
			Xunit.Assert.Equal
				(expected:Role.Administrator,
				actual: result.Value);
			// **************************************************

			// **************************************************
			Xunit.Assert.Empty(result.Errors);
			Xunit.Assert.Empty(result.Successes);
			// **************************************************
		}

		[Xunit.Fact]
		public void Add_Programmer_Role()
		{
			var result =
				Role.GetByValue(value: 4);

			// **************************************************
			Xunit.Assert.False(condition: result.IsFailed);
			Xunit.Assert.True(condition: result.IsSuccess);
			// **************************************************

			// **************************************************
			Xunit.Assert.Equal
				(expected: Role.Programmer,
				actual: result.Value);
			// **************************************************

			// **************************************************
			Xunit.Assert.Empty(result.Errors);
			Xunit.Assert.Empty(result.Successes);
			// **************************************************
		}
	}
}
